requirements.txt
torch==2.0.1+cu118
torchvision==0.15.2+cu118
segmentation-models-pytorch==0.3.0
albumentations==1.3.1
onnx==1.14.0
onnxruntime-gpu==1.15.1
tensorrt==8.6.1
pycuda==2023.1
numpy==1.24.3
opencv-python==4.7.0.72
matplotlib==3.7.1
tqdm==4.65.0
scikit-learn==1.2.2